requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
huggingface-hub==0.32.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
packaging==25.0
pillow==11.2.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==80.9.0
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
